Refactor quadbin functions into quadbin-utils
