+++ modules/carto/src/layers/quadbin-utils.ts
@@ -1,11 +1,104 @@
 import {worldToLngLat} from '@math.gl/web-mercator';
-import {quadbinToTile, tileToQuadkey} from './quadbin-tileset-2d';
 
 const TILE_SIZE = 512;
 
+const B = [
+  0x5555555555555555n,
+  0x3333333333333333n,
+  0x0f0f0f0f0f0f0f0fn,
+  0x00ff00ff00ff00ffn,
+  0x0000ffff0000ffffn,
+  0x00000000ffffffffn
+];
+const S = [0n, 1n, 2n, 4n, 8n, 16n];
+
+function indexToBigInt(index: string): bigint {
+  return BigInt(`0x${index}`);
+}
+
+function bigIntToIndex(quadbin: bigint): string {
+  return quadbin.toString(16);
+}
+
+export function tileToQuadbin(tile): string {
+  if (tile.z < 0 || tile.z > 26) {
+    throw new Error('Wrong zoom');
+  }
+  const z = BigInt(tile.z);
+  let x = BigInt(tile.x) << (32n - z);
+  let y = BigInt(tile.y) << (32n - z);
+
+  for (let i = 0; i < 5; i++) {
+    const s = S[5 - i];
+    const b = B[4 - i];
+    x = (x | (x << s)) & b;
+    y = (y | (y << s)) & b;
+  }
+
+  const quadbin =
+    0x4000000000000000n |
+    (1n << 59n) | // | (mode << 59) | (mode_dep << 57)
+    (z << 52n) |
+    ((x | (y << 1n)) >> 12n) |
+    (0xfffffffffffffn >> (z * 2n));
+  return bigIntToIndex(quadbin);
+}
+
+export function quadbinToTile(index: string) {
+  const quadbin = indexToBigInt(index);
+  const mode = (quadbin >> 59n) & 7n;
+  const modeDep = (quadbin >> 57n) & 3n;
+  const z = (quadbin >> 52n) & 0x1fn;
+  const q = (quadbin & 0xfffffffffffffn) << 12n;
+
+  if (mode !== 1n && modeDep !== 0n) {
+    throw new Error('Wrong mode');
+  }
+
+  let x = q;
+  let y = q >> 1n;
+
+  for (let i = 0; i < 6; i++) {
+    const s = S[i];
+    const b = B[i];
+    x = (x | (x >> s)) & b;
+    y = (y | (y >> s)) & b;
+  }
+
+  x = x >> (32n - z);
+  y = y >> (32n - z);
+
+  return {z: Number(z), x: Number(x), y: Number(y)};
+}
+
+export function quadbinZoom(index: string) {
+  const quadbin = indexToBigInt(index);
+  return (quadbin >> 52n) & 0x1fn;
+}
+
+export function quadbinParent(index: string) {
+  const quadbin = indexToBigInt(index);
+  const zparent = quadbinZoom(index) - 1n;
+  const parent =
+    (quadbin & ~(0x1fn << 52n)) | (zparent << 52n) | (0xfffffffffffffn >> (zparent * 2n));
+  return bigIntToIndex(parent);
+}
+
+export function tileToQuadkey(tile) {
+  let index = '';
+  for (let z = tile.z; z > 0; z--) {
+    let b = 0;
+    const mask = 1 << (z - 1);
+    if ((tile.x & mask) !== 0) b++;
+    if ((tile.y & mask) !== 0) b += 2;
+    index += b.toString();
+  }
+  return index;
+}
+
 // Use existing functions for now
 export function quadbinToWorldBounds(quadbin: string): [number[], number[]] {
-  const tile = quadbinToTile({i: quadbin});
+  const tile = quadbinToTile(quadbin);
   const quadkey = tileToQuadkey(tile);
   return quadkeyToWorldBounds(quadkey);
 }
